+++ test/test_helper.exs
@@ -1,1 +1,1 @@
-ExUnit.start(exclude: [:integration, :fidelity, :private_fixture, :ddl_postgres])
+ExUnit.start(exclude: [:integration, :fidelity, :private_fixture, :ddl_postgres, :syntax_check])